refactor(slack-app): run queued messages as child workflows
The conversation queue now executes each message as a child
PostHogCodeSlackMentionWorkflow under the per-message workflow ID,
instead of driving the extracted process_mention_message orchestration
inline. This reverts the helpers extraction: the per-message workflow
returns to its exact pre-branch form and stays the single processing
implementation, so flag-off behavior is identical to master by
construction rather than by review.

What this buys:
- Interactive signals (repo picker, authorship) route straight to the
  child, which bakes its own ID into the prompts it posts — the
  conversation workflow no longer hosts those handlers, and the stale
  picker cross-message edge case is gone.
- slack_mention_workflow_id persisted by task creation points at a real
  execution again, so the debug-tool Temporal link works under the flag.
- Parent history stays small (a few events per message), leaving
  continue_as_new as a rarely needed safety valve.

Child starts use ALLOW_DUPLICATE (mirroring the standalone dispatch).
The default parent close policy (terminate) is kept deliberately:
killing the queue means stopping the conversation, in-flight child
included; normal exits never have a child in flight.
